refactor: updated some issues
