+++ src/components/Main/Main.tsx
@@ -1,4 +1,4 @@
-import type { IPokemon } from '../../src/utils/types.ts';
+import type { IPokemon } from '../../utils/types.ts';
 import { Component } from 'react';
 import CardList from '../CardList/CardList.tsx';
 import Skeleton from '../Skeleton/Skeleton.tsx';
@@ -15,25 +15,35 @@ interface Props {
 
 export class Main extends Component<Props> {
   render() {
-    const { data, isLoading, onLoadMore, hasNext } = this.props;
+    const { data, isLoading, onLoadMore, hasNext, error } = this.props;
+
+    let content;
+    if (error) {
+      content = (
+        <div className="text-red-600 font-semibold bg-red-100 p-4 rounded text-center">
+          {error}
+        </div>
+      );
+    } else if (isLoading) {
+      content = (
+        <div className="flex flex-col items-center gap-4">
+          <Skeleton />
+          <Loader />
+        </div>
+      );
+    } else if (data.length === 0) {
+      content = (
+        <div className="text-center text-gray-600 py-8">No results found.</div>
+      );
+    } else {
+      content = <CardList data={data} />;
+    }
 
     return (
       <main className="border rounded-lg p-4 shadow-sm bg-red-400">
         <section className="border rounded-lg p-4 bg-white shadow">
           <h2 className="text-lg font-semibold mb-4">Results</h2>
-
-          {this.props.error ? (
-            <div className="text-red-600 font-semibold bg-red-100 p-4 rounded text-center">
-              {this.props.error}
-            </div>
-          ) : isLoading ? (
-            <div className="flex flex-col items-center gap-4">
-              <Skeleton />
-              <Loader />
-            </div>
-          ) : (
-            <CardList data={data} />
-          )}
+          {content}
         </section>
 
         {hasNext && !isLoading && (
