+++ src/main.tsx
@@ -2,8 +2,8 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
-import { ErrorBoundary } from '../components/ErrorBoundary/ErrorBoundary.tsx';
-import { FallbackUi } from '../components/FallbackUI/FallbackUi.tsx';
+import { ErrorBoundary } from './components/ErrorBoundary/ErrorBoundary.tsx';
+import { FallbackUi } from './components/FallbackUI/FallbackUi.tsx';
 
 const rootElement = document.getElementById('root');
 if (!rootElement) throw new Error('Root element not found');
